backend/service_projects/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.36
pydantic==2.9.2
httpx==0.27.2




